refactor wallet hook
